+++ lapdog/__init__.py
@@ -1,3 +1,3 @@
 from dalmatian import *
 from .lapdog import *
-__version__ = "0.14.0"
+__version__ = "0.14.1"